+++ src/Main.gd
@@ -121,8 +121,8 @@ some useful [b][SYSTEM OPTIONS][/b] are:
 	static func enable_export(_project: Project, _next_arg: String):
 		return true
 
-	static func enable_spritesheet(_project: Project, _next_arg: String):
-		Export.current_tab = Export.ExportTab.SPRITESHEET
+	static func enable_spritesheet(project: Project, _next_arg: String):
+		project.export_tab = Export.ExportTab.SPRITESHEET
 		return true
 
 	static func set_output(project: Project, next_arg: String) -> void:
@@ -134,10 +134,10 @@ some useful [b][SYSTEM OPTIONS][/b] are:
 			var extension := next_arg.get_extension()
 			project.file_format = Export.get_file_format_from_extension(extension)
 
-	static func set_export_scale(_project: Project, next_arg: String) -> void:
+	static func set_export_scale(project: Project, next_arg: String) -> void:
 		if not next_arg.is_empty():
 			if next_arg.is_valid_float():
-				Export.resize = next_arg.to_float() * 100
+				project.export_resize = next_arg.to_float() * 100
 
 	static func set_frames(project: Project, next_arg: String) -> void:
 		if not next_arg.is_empty():
@@ -156,36 +156,36 @@ some useful [b][SYSTEM OPTIONS][/b] are:
 					for frame in range(frame_number_1, frame_number_2 + 1):
 						project.selected_cels.append([frame, project.current_layer])
 						project.change_cel(frame)
-						Export.frame_current_tag = Export.ExportFrames.SELECTED_FRAMES
+						project.export_frame_current_tag = Export.ExportFrames.SELECTED_FRAMES
 			elif next_arg.is_valid_int():
 				var frame_number := next_arg.to_int() - 1
 				frame_number = clampi(frame_number, 0, project.frames.size() - 1)
 				project.selected_cels = [[frame_number, project.current_layer]]
 				project.change_cel(frame_number)
-				Export.frame_current_tag = Export.ExportFrames.SELECTED_FRAMES
+				project.export_frame_current_tag = Export.ExportFrames.SELECTED_FRAMES
 
-	static func set_direction(_project: Project, next_arg: String) -> void:
+	static func set_direction(project: Project, next_arg: String) -> void:
 		if not next_arg.is_empty():
 			next_arg = next_arg.to_lower()
 			if next_arg == "0" or next_arg.contains("forward"):
-				Export.direction = Export.AnimationDirection.FORWARD
+				project.export_direction = Export.AnimationDirection.FORWARD
 			elif next_arg == "1" or next_arg.contains("backward"):
-				Export.direction = Export.AnimationDirection.BACKWARDS
+				project.export_direction = Export.AnimationDirection.BACKWARDS
 			elif next_arg == "2" or next_arg.contains("ping"):
-				Export.direction = Export.AnimationDirection.PING_PONG
+				project.export_direction = Export.AnimationDirection.PING_PONG
 			else:
-				print(Export.AnimationDirection.keys()[Export.direction])
+				print(Export.AnimationDirection.keys()[project.export_direction])
 		else:
-			print(Export.AnimationDirection.keys()[Export.direction])
+			print(Export.AnimationDirection.keys()[project.export_direction])
 
-	static func set_json(_project: Project, _next_arg: String) -> void:
-		Export.export_json = true
+	static func set_json(project: Project, _next_arg: String) -> void:
+		project.export_json = true
 
-	static func set_split_layers(_project: Project, _next_arg: String) -> void:
-		Export.split_layers = true
+	static func set_split_layers(project: Project, _next_arg: String) -> void:
+		project.export_split_layers = true
 
-	static func set_sheet_layers_as_separate_files(_project: Project, _next_arg: String) -> void:
-		Export.sheet_layers_as_separate_files = true
+	static func set_sheet_layers_as_separate_files(project: Project, _next_arg: String) -> void:
+		project.export_sheet_layers_as_separate_files = true
 
 	static func dummy(_project: Project, _next_arg: String) -> void:
 		pass